requirements.txt
certifi==2018.4.16
chardet==3.0.4
click==6.7
dateutils==0.6.6
Flask==1.0.2
Flask-Cors==3.0.6
idna==2.7
itsdangerous==0.24
Jinja2==2.10
lxml==4.2.3
MarkupSafe==1.0
numpy==1.14.5
pandas==0.23.2
python-dateutil==2.7.3
pytz==2018.5
requests==2.19.1
six==1.11.0
urllib3==1.23
Werkzeug==0.14.1
